refactor(auth): migrate Google OAuth endpoints to dedicated module
