+++ app/core/config.py
@@ -246,6 +246,15 @@ class ServerSettings(BaseSettings):
     )
 
 
+class FrontEndSettings(BaseSettings):
+    FRONTEND_URL: str = Field(default="http://localhost:3000", alias="FRONTEND_URL")
+    model_config = SettingsConfigDict(
+        env_file=".env",
+        case_sensitive=True,
+        extra="ignore",
+    )
+
+
 class Settings(BaseSettings):
     """SRE Bot configuration settings."""
 
@@ -255,6 +264,7 @@ class Settings(BaseSettings):
 
     # Server settings
     server: ServerSettings
+    frontend: FrontEndSettings
 
     # Integration settings
     slack: SlackSettings
@@ -281,6 +291,7 @@ class Settings(BaseSettings):
     def __init__(self, **kwargs):
         settings_map = {
             "server": ServerSettings,
+            "frontend": FrontEndSettings,
             "slack": SlackSettings,
             "aws": AwsSettings,
             "google_workspace": GoogleWorkspaceSettings,
