+++ app/tests/server/test_server.py
@@ -488,170 +488,6 @@ def test_react_app():
     assert "text/html" in response.headers["content-type"]
 
 
-# Test the logout endpoint
-def test_logout_endpoint():
-    # Test that the endpoint returns a 200 status code
-    response = client.get("/logout")
-    assert response.status_code == 200
-
-    # Test that the user session is removed
-    response = client.get("/home")
-    assert response.status_code == 200
-    assert "user" not in response.cookies
-
-
-# Test the login endpoint and that it redirects to the Google OAuth page
-def test_login_endpoint():
-    response = client.get("/login")
-    assert response.status_code == 200
-    assert "https://accounts.google.com/o/oauth2/v2/auth" in str(response.url)
-
-
-# Test the login endpoint converts the redirect_uri to https
-@patch("server.server.settings")
-def test_login_endpoint_redirect_uri_prod(mock_settings):
-    # Make a test request to the login endpoint
-    mock_settings.is_production = True
-    response = client.get("/login")
-
-    # assert the call is successful
-    assert response.status_code == 200
-
-    # Set up the expected redirect_uri
-    redirect_uri = urllib.parse.quote_plus("http://testserver/auth")
-    # Convert to https for production
-    redirect_uri = redirect_uri.__str__().replace("http", "https")
-
-    # assert that the response url we get from the login endpoint contains the redirect_uri replaced with https
-    assert response.url.__str__().__contains__("redirect_uri=" + redirect_uri)
-
-
-# Test the login endpoing that does not convert the redirect uri
-@patch("server.server.settings")
-def test_login_endpoint_redirect_uri_dev(mock_settings):
-    # Setup the mock to simulate dev environment
-    mock_settings.is_production = False
-
-    # Make a test request to the login endpoint
-    response = client.get("/login")
-
-    # assert the call is successful
-    assert response.status_code == 200
-
-    # Set up the expected redirect_uri (without https conversion for dev)
-    redirect_uri = urllib.parse.quote_plus("http://testserver/auth")
-
-    # assert that the response url we get from the login endpoint contains the redirect_uri is not replaced with https
-    assert response.url.__str__().__contains__("redirect_uri=" + redirect_uri)
-
-
-# Test the auth endpoint
-def test_auth_endpoint():
-    response = client.get("/auth")
-    assert response.status_code == 200
-    assert "http://testserver/auth" in str(response.url)
-
-
-# Test the user endpoint, logged in
-def test_user_route_logged_in():
-    # Simulate a logged-in session by creating a mock request with session data
-    session_data = {"user": {"given_name": "FirstName"}}
-    headers = {"Cookie": f"session={session_data}"}
-    response = client.get("/user", headers=headers)
-    assert response.status_code == 200
-
-
-# Test the user endpoing, not logged in
-def test_user_endpoint_with_no_logged_in_user():
-    response = client.get("/user")
-    assert response.status_code == 200
-    assert response.json() == {"error": "Not logged in"}
-
-
-@pytest.mark.asyncio
-async def test_logout_rate_limiting():
-    # Create a custom transport to mount the ASGI app
-    transport = httpx.ASGITransport(app=app)
-
-    # Use the MockTransport in the AsyncClient
-    async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
-        # Make 5 requests to the logout endpoint
-        for _ in range(5):
-            response = await client.get("/logout")
-            assert response.status_code == 307
-            assert response.url.path == "/logout"
-
-        # The 6th request should be rate limited
-        response = await client.get("/logout")
-        assert response.status_code == 429
-        assert response.json() == {"message": "Rate limit exceeded"}
-
-
-@pytest.mark.asyncio
-async def test_login_rate_limiting():
-    # Create a custom transport to mount the ASGI app
-    transport = httpx.ASGITransport(app=app)
-
-    async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
-        # Set the environment variable for the test
-        os.environ["ENVIRONMENT"] = "dev"
-
-        # Make 5 requests to the login endpoint
-        for _ in range(5):
-            response = await client.get("/login")
-            assert response.status_code == 302
-
-        # The 6th request should be rate limited
-        response = await client.get("/login")
-        assert response.status_code == 429
-        assert response.json() == {"message": "Rate limit exceeded"}
-
-
-@pytest.mark.asyncio
-async def test_auth_rate_limiting():
-    # Create a custom transport to mount the ASGI app
-    transport = httpx.ASGITransport(app=app)
-
-    async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
-        # Mock the OAuth process
-        with patch(
-            "server.server.oauth.google.authorize_access_token", new_callable=AsyncMock
-        ) as mock_auth:
-            mock_auth.return_value = {
-                "userinfo": {"name": "Test User", "email": "test@test.com"}
-            }
-
-            # Make 5 requests to the auth endpoint
-            for _ in range(5):
-                response = await client.get("/auth")
-                assert response.status_code == 307
-
-            # The 6th request should be rate limited
-            response = await client.get("/auth")
-            assert response.status_code == 429
-            assert response.json() == {"message": "Rate limit exceeded"}
-
-
-@pytest.mark.asyncio
-async def test_user_rate_limiting():
-    # Create a custom transport to mount the ASGI app
-    transport = httpx.ASGITransport(app=app)
-
-    async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
-        # Simulate a logged in session
-        session_data = {"user": {"given_name": "FirstName", "email": "test@test.com"}}
-        headers = {"Cookie": f"session={session_data}"}
-        # Make 10 requests to the user endpoint
-        for _ in range(10):
-            response = await client.get("/user", headers=headers)
-            assert response.status_code == 200
-
-        # The 11th request should be rate limited
-        response = await client.get("/user", headers=headers)
-        assert response.status_code == 429
-        assert response.json() == {"message": "Rate limit exceeded"}
-
-
 @patch(
     "server.server.webhooks.get_webhook",
     return_value={"channel": {"S": "test-channel"}},
